share modal refactor
